Fits.Py codes from May 17, 2022
Started using the Cauchy loss for the optimize.least_squares, which yielded better results for the reflective and background constants, as well as reasonable values for b and c, except for one c value, which was set to the bounds. I'm currently investigating what can be done to solve this.